Moving the frontend into src/frontend
