+++ src/frontend/pages/WishListPage.tsx
@@ -1,12 +1,7 @@
 import * as stylex from "@stylexjs/stylex"
 import { useState } from "react"
 import { useParams } from "react-router-dom"
-// import { Button } from "src/assets/Button"
-// import { MenuBar } from "src/assets/MenuBar"
-// import { PopUp } from "src/assets/PopUp"
-// import { NewWishItemForm } from "src/components/wishes/NewWishItemForm"
-// import { WishList } from "src/components/wishList/WishList"
-import { WishListType } from "src/types"
+import { WishListType } from "src/frontend/types"
 import { Button } from "../assets/Button"
 import { MenuBar } from "../assets/MenuBar"
 import { PopUp } from "../assets/PopUp"
